Database Merger
This script is designed to merge two SQLite3 databases together using recovered data, .sql files, and a .json file. This is incredibly specific, but it does "work on my machine", and should hopefully work on Clueless' side, too.

There is a bug with some user stats "progress" grouped by canvas speed being incorrect, however I currently do not have a solution for this. The good thing is this can be run again at a later date.

To enable this merger in full, you will need:
- The old database as database_malformed.db
- The new database as database_new.db
- An exported version of the "record" table as record.sql
- An exported version of both "command_usage" tables, from the new and old databases, as both malformed_command_usage.sql & new_command_usage.sql
- A .json of all data available from c78a as 78a.json
- A bit of patience, this does take a bit